+++ cmd/hidden_lake/service/internal/handler/handler_test.go
@@ -206,7 +206,7 @@ func testNewNode(dbPath, addr string) anonymity.INode {
 		),
 		db,
 		testNewNetworkNode(addr),
-		queue.NewMessageQueue(
+		queue.NewMessageQueueProcessor(
 			queue.NewSettings(&queue.SSettings{
 				FNetworkMask:      networkMask,
 				FWorkSizeBits:     testutils.TCWorkSize,