+++ pkg/network/anonymity/logger/types.go
@@ -35,7 +35,6 @@ const (
 	CLogWarnIncorrectResponse
 
 	// ERRO
-	CLogErroEncryptPayload
 	CLogErroDatabaseGet
 	CLogErroDatabaseSet
 )